Fix #65 - coerce domain to date.
Previously it was possible to set a time scale’s domain to valid numbers but not
valid dates: the allowable range for Number is much larger than the allowable
range for Date. (The latter is only ±8.64e15!) By coercing the domain to date,
we avoid a nearly-infinite loop when trying to compute ticks.